backend/pyproject.toml
[project]
name = "tech-radar"
version = "0.1.0"
description = "FastAPI backend for tech-radar"
requires-python = ">=3.11,<3.12"
readme = "README.md"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
	"E",  # pycodestyle errors
	"F",  # pyflakes
	"I",  # isort
	"B",  # flake8-bugbear
	"UP", # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ['pydantic.mypy']
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
no_implicit_optional = true
check_untyped_defs = true
explicit_package_bases = true
namespace_packages = true
mypy_path = ["."]

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "tech-radar"
version = "0.1.0"
description = "FastAPI backend for tech-radar"
readme = "README.md"
authors = ["Your Name <you@example.com>"]
packages = []

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = ">=0.118.0"
uvicorn = { version = ">=0.37.0", extras = ["standard"] }
beanie = "^2.0.0"
pydantic-settings = "^2.11.0"
pydantic = "^2.11.9"
pymongo = "^4.15.2"

[tool.poetry.group.dev.dependencies]
ruff = ">=0.13.2"
mypy = ">=1.18.2"
types-requests = ">=2.32.0.20240914"
fastapi-cli = ">=0.0.5"
pytest = "^8.0.0"
pytest-asyncio = "^1.2.0"
httpx = "^0.28.1"
mongomock-motor = "^0.0.36"
pytest-mock = "^3.14.0"

[tool.pytest.ini_options]
filterwarnings = [
	# Beanie is raising deprecation warnings which I can't control
	"ignore:Accessing the 'model_fields' attribute on the instance is deprecated",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
